refactor: remove standalone voice user subscription
