+++ bigbluebutton-html5/imports/ui/components/video-provider/video-list/video-list-item/container.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import useCurrentUser from '/imports/ui/core/hooks/useCurrentUser';
 import { layoutSelect, layoutDispatch } from '/imports/ui/components/layout/context';
@@ -8,7 +8,6 @@ import { Layout } from '/imports/ui/components/layout/layoutTypes';
 import useSettings from '/imports/ui/services/settings/hooks/useSettings';
 import { SETTINGS } from '/imports/ui/services/settings/enums';
 import { useStorageKey } from '/imports/ui/services/storage/hooks';
-import useVoiceUsers from '/imports/ui/components/audio/audio-graphql/hooks/useVoiceUsers';
 import useWhoIsTalking from '/imports/ui/core/hooks/useWhoIsTalking';
 import useWhoIsUnmuted from '/imports/ui/core/hooks/useWhoIsUnmuted';
 
@@ -56,19 +55,13 @@ const VideoListItemContainer: React.FC<VideoListItemContainerProps> = (props) =>
   const amIModerator = currentUserData?.isModerator;
 
   const disabledCams = useStorageKey('disabledCams') || [];
-  const voiceUsers = useVoiceUsers((v) => ({
-    listenOnly: v.listenOnly,
-    joined: v.joined,
-    userId: v.userId,
-  }));
   const { data: talkingUsers } = useWhoIsTalking();
   const { data: unmutedUsers } = useWhoIsUnmuted();
-  const voiceUser = voiceUsers.data?.find((v) => v.userId === userId) ?? {};
-  const voiceData = useMemo(() => ({
-    ...voiceUser,
+  const voiceUser = stream.type !== 'connecting' && stream.voice ? {
+    ...stream.voice,
     talking: talkingUsers[userId],
     muted: !unmutedUsers[userId],
-  }), [voiceUser, talkingUsers, unmutedUsers, userId]);
+  } : {};
 
   return (
     <VideoListItem
@@ -90,7 +83,7 @@ const VideoListItemContainer: React.FC<VideoListItemContainerProps> = (props) =>
       onVirtualBgDrop={onVirtualBgDrop}
       settingsSelfViewDisable={settingsSelfViewDisable}
       stream={stream}
-      voiceUser={voiceData}
+      voiceUser={voiceUser}
     />
   );
 };
